requirements.txt
python-dotenv
lnhistoryclient
pyzmq
psycopg[binary]
psycopg_pool
crc32c